feat: Placed affirm into a new cog called Message
