+++ cogs/messages/affirmations.txt
